Add files via upload
I am trying to write a function that prompts a user for an input from 20 to 100, then write a txt file with that many numbers in it.  I will than have to read those inputs into a series of singly linked lists (in the order they appear, in reverse order, highest to lowest, then lowest to highest).  Any help would be appriciated.